Save client VK when setting entity
It may be the case that a library creates a client and passes it to
another piece of code, which may need the VK. Because the VK is only
ever displayed when the client is created, the user would need to know
to provide both pieces of information: the client instance AND the vk.
By caching the vk in the client object, we remove that need.

One caveat with this is that I'm not sure where the async entity methods
actually return/display/parse the VK, so that would need to be done