Allow callbacks to execute bosswave functions
Currently, executing a bosswave function from within a callback blocks
(such as when we want to publish a message whenever we receive a trigger
on a subscription). I think this was caused by some deadlock on
the socket to the client. To get around this, I start up a second thread
and have the callbacks executed there. Done some basic testing and this
seems to work fine